+++ examples/src/main/java/org/partiql/examples/CSVJavaExample.java
@@ -3,13 +3,7 @@ package org.partiql.examples;
 import org.jetbrains.annotations.NotNull;
 import org.partiql.examples.util.Example;
 import org.partiql.lang.CompilerPipeline;
-import org.partiql.lang.eval.BaseExprValue;
-import org.partiql.lang.eval.Bindings;
-import org.partiql.lang.eval.EvaluationSession;
-import org.partiql.lang.eval.ExprValue;
-import org.partiql.lang.eval.ExprValueExtensionsKt;
-import org.partiql.lang.eval.ExprValueType;
-import org.partiql.lang.eval.Expression;
+import org.partiql.lang.eval.*;
 
 import java.io.PrintStream;
 import java.util.Arrays;