+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/NaturalExprValueComparators.kt
@@ -16,6 +16,7 @@ package org.partiql.lang.eval
 
 import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.eval.NaturalExprValueComparators.NullOrder
+import org.partiql.lang.eval.NaturalExprValueComparators.Order
 import org.partiql.lang.util.compareTo
 import org.partiql.lang.util.isNaN
 import org.partiql.lang.util.isNegInf