+++ partiql-lang/src/test/kotlin/org/partiql/lang/errors/ParserErrorsTest.kt
@@ -19,8 +19,8 @@ package org.partiql.lang.errors
 import com.amazon.ion.Timestamp
 import org.junit.Test
 import org.partiql.lang.syntax.PartiQLParserTestBase
-import org.partiql.lang.syntax.antlr.PartiQLParser
 import org.partiql.lang.util.getAntlrDisplayString
+import org.partiql.parser.antlr.PartiQLParser
 
 class ParserErrorsTest : PartiQLParserTestBase() {